feat: refactor the era5 to reverse the order
