+++ src/geodata/datasets/era5/wind_3d/__init__.py
@@ -13,6 +13,7 @@
 # You should have received a copy of the GNU General Public License
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-from .wind_solar import ERA5WindSolarMonthlyDataset
+from .hourly import ERA5Wind3DHourlyDataset
 
-__all__ = ["ERA5WindSolarMonthlyDataset"]
+__all__ = ["ERA5Wind3DHourlyDataset"]
+
