+++ src/geodata/datasets/era5/wind_3d/hourly.py
@@ -20,8 +20,8 @@ from pathlib import Path
 
 import xarray as xr
 
-from ..._base import AtomicDataset
-from .._base import ERA5BaseDataset
+from ...._base import AtomicDataset
+from ..._base import ERA5BaseDataset
 
 logger = logging.getLogger(__name__)
 
@@ -114,3 +114,4 @@ class ERA5Wind3DHourlyDataset(ERA5BaseDataset):
                 ds.to_netcdf(save_path)
 
         logger.info("File downloaded: %s", save_path)
+
