+++ src/geodata/datasets/era5/wind_solar/__init__.py
@@ -13,7 +13,8 @@
 # You should have received a copy of the GNU General Public License
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-from .wind_3d import ERA5Wind3DHourlyDataset
-from .wind_solar import ERA5WindSolarHourlyDataset
+from .hourly import ERA5WindSolarHourlyDataset
+from .monthly import ERA5WindSolarMonthlyDataset
 
-__all__ = ["ERA5WindSolarHourlyDataset", "ERA5Wind3DHourlyDataset"]
+__all__ = ["ERA5WindSolarHourlyDataset", "ERA5WindSolarMonthlyDataset"]
+
