+++ src/geodata/datasets/era5/wind_solar/hourly.py
@@ -22,8 +22,8 @@ from pathlib import Path
 
 import xarray as xr
 
-from ..._base import AtomicDataset
-from .._base import ERA5BaseDataset
+from ...._base import AtomicDataset
+from ..._base import ERA5BaseDataset
 
 logger = logging.getLogger(__name__)
 
@@ -130,3 +130,4 @@ class ERA5WindSolarHourlyDataset(ERA5BaseDataset):
 
                 logger.info("Preprocessing complete with zipfile")
                 logger.info("Successfully downloaded to %s", save_path)
+
